import - export, question bank nha

+++ app/Http/Controllers/QuizController.php
@@ -4,8 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Models\CreateQuizModel;
 use App\Models\QuizModel;
+use App\Models\SpreadsheetModel;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use PhpOffice\PhpSpreadsheet\IOFactory;
+use PhpOffice\PhpSpreadsheet\Style\Alignment;
+use PhpOffice\PhpSpreadsheet\Style\Border;
 
 class QuizController extends Controller
 {
@@ -14,12 +18,13 @@ class QuizController extends Controller
         $model = new QuizModel();
 //        dd($model->getInfo());
         return view('auth.quiz.quiz',
-            ['data'=>$model->getInfo()]);
+            ['data' => $model->getInfo()]);
     }
 
-    function getViewQuestionBank() {
+    function getViewQuestionBank()
+    {
         $model = new QuizModel();
-        return view('auth.quiz.question-bank',['courses' => $model->getCourse()]);
+        return view('auth.quiz.question-bank', ['courses' => $model->getCourse()]);
     }
 
     function addQuestion(Request $request)
@@ -55,7 +60,7 @@ class QuizController extends Controller
 
         return response()->json([
             'success' => true,
-            'status'  => 200,
+            'status' => 200,
             'message' => 'Question added successfully',
         ]);
     }
@@ -63,7 +68,7 @@ class QuizController extends Controller
     function getQuestionList($id)
     {
         $question_list = DB::table('question_bank')
-            ->join('courses','question_bank.course_id','=','courses.course_id')
+            ->join('courses', 'question_bank.course_id', '=', 'courses.course_id')
             ->where('question_bank.course_id', $id)
             ->get();
         return response()->json([
@@ -76,11 +81,18 @@ class QuizController extends Controller
 //        dd($id);
         $question = QuizModel::findOrFail($id);
 
+        // Define the directory path
         $directoryPath = public_path('question_bank_image/');
-        $filePath = $directoryPath . $question->question_image;
 
-        if (file_exists($filePath)) {
-            unlink($filePath);
+        // Check if the question_image is not null
+        if (!is_null($question->question_image)) {
+            // Define the file path
+            $filePath = $directoryPath . $question->question_image;
+
+            // Check if the file exists before attempting to delete it
+            if (file_exists($filePath)) {
+                unlink($filePath);
+            }
         }
 
         $question->delete();
@@ -135,4 +147,95 @@ class QuizController extends Controller
             'question' => $question,
         ]);
     }
+
+    public function import(Request $request, $id)
+    {
+        try {
+            $dataExcel = SpreadsheetModel::readExcel($request->file('file-excel'));
+            $num_row = 0;
+            $dataArray = array();
+//            dd($dataExcel['data']);
+            foreach ($dataExcel['data'] as $item) {
+                $num_row++;
+                if ($num_row == 1) {
+                    continue; // Skip header row
+                }
+
+                $data = [
+                    'question' => trim($item[1]),
+                    'question_a' => trim($item[2]),
+                    'question_b' => trim($item[3]),
+                    'question_c' => trim($item[4]),
+                    'question_d' => trim($item[5]),
+                    'correct' => trim($item[6]),
+                    'course_id' => $id,
+                ];
+
+                $dataArray[] = $data;
+            }
+
+            if (!empty($dataArray)) {
+                foreach ($dataArray as $data) {
+                    QuizModel::create($data);
+                }
+                return response()->json(['status' => 200, 'message' => 'Import successfully']);
+            } else {
+                return response()->json(['status' => 400, 'message' => 'No data for import']);
+            }
+        } catch (\Exception $e) {
+            return response()->json(['status' => 500, 'message' => 'An error occurred ' . $e->getMessage()]);
+        }
+
+    }
+
+    public function export($id)
+    {
+        $inputFileName = public_path('excel-example/question-bank-export.xlsx');
+        $inputFileType = IOFactory::identify($inputFileName);
+        $objReader = IOFactory::createReader($inputFileType);
+        $excel = $objReader->load($inputFileName);
+
+        $excel->setActiveSheetIndex(0);
+        $excel->getDefaultStyle()->getFont()->setName('Times New Roman');
+
+        $stt = 1;
+        $cell = $excel->getActiveSheet();
+        $question_bank = DB::table('question_bank')->where('course_id', $id)->get();
+        $num_row = 2;
+        foreach ($question_bank as $row) {
+            $cell->setCellValue('A' . $num_row, $stt++);
+            $cell->setCellValue('B' . $num_row, $row->question);
+            $cell->setCellValue('C' . $num_row, $row->question_a);
+            $cell->setCellValue('D' . $num_row, $row->question_b);
+            $cell->setCellValue('E' . $num_row, $row->question_c);
+            $cell->setCellValue('F' . $num_row, $row->question_d);
+            $cell->setCellValue('G' . $num_row, $row->correct);
+            $borderStyle = $cell->getStyle('A' . $num_row . ':G' . $num_row)->getBorders();
+            $borderStyle->getAllBorders()->setBorderStyle(Border::BORDER_THIN);
+            $cell->getStyle('A' . $num_row . ':G' . $num_row)->getAlignment()->setHorizontal(Alignment::HORIZONTAL_LEFT);
+            $num_row++;
+        }
+        foreach (range('A', 'G') as $columnID) {
+            $excel->getActiveSheet()->getColumnDimension($columnID)->setAutoSize(true);
+        }
+
+        $fileName = "Question-Bank-Export-" . time() . ".xlsx";
+        $filePath = public_path('excel-download/' . $fileName);
+
+        $writer = IOFactory::createWriter($excel, 'Xlsx');
+        $writer->save($filePath);
+
+        return response()->json(['success' => true, 'status' => 200, 'message' => 'Export successfully', 'file' => 'excel-download/' . $fileName]);
+    }
+
+    public function deleteExportedFile(Request $request)
+    {
+        $filePath = public_path($request->input('file'));
+        if (file_exists($filePath)) {
+            unlink($filePath);
+            return response()->json(['success' => true, 'message' => 'File deleted successfully']);
+        }
+        return response()->json(['success' => false, 'message' => 'File not found']);
+    }
+
 }
